Convert index.js to index.tsx
