+++ client/src/App.tsx
@@ -10,7 +10,6 @@ function App() {
         <p>
           Edit <code>src/App.tsx</code> and save to reload.
         </p>
-        <div>Please work</div>
         <a
           className="App-link"
           href="https://reactjs.org"
